[#DCM-146]: Fix Eclipse warnings  * Remove unnecessary SuppressWarnings annotations  * Add missing @Override annotations  * Fix 'empty block should be documented'  * Remove unnecessary casts  * Remove unnecessary imports  * Add @Deprecated annotations

+++ dcm4che2/trunk/dcm4che-core/src/test/java/org/dcm4che2/io/DicomOutputStreamTest.java
@@ -171,7 +171,7 @@ public class DicomOutputStreamTest extends TestCase {
     }
     
     private byte[] loadBytes(InputStream in) throws IOException {
-        int remain = (int) in.available();
+        int remain = in.available();
         byte[] b = new byte[remain];
         try {
             while (remain > 0) {
